Coap Server and Client package merged to one package node-wot-protocol-coap
