+++ packages/integration-tests/package.json
@@ -16,8 +16,7 @@
     "mocha": "3.5.3",
     "mocha-typescript": "1.1.8",
     "node-wot-protocols": "0.3.0-SNAPSHOT.2",
-    "node-wot-protocols-coap-client": "0.3.0-SNAPSHOT.2",
-    "node-wot-protocols-coap-server": "0.3.0-SNAPSHOT.2",
+    "node-wot-protocol-coap": "0.3.0-SNAPSHOT.2",
     "node-wot-protocols-http-client": "0.3.0-SNAPSHOT.2",
     "node-wot-protocols-http-server": "0.3.0-SNAPSHOT.2",
     "node-wot": "0.3.0-SNAPSHOT.2",
