+++ packages/integration-tests/test/ProtocolsTest.ts
@@ -8,7 +8,7 @@ import { expect, should } from "chai";
 should();
 
 import HttpServer from "node-wot-protocols-http-server";
-import CoapServer from "node-wot-protocols-coap-server"; 
+import {CoapServer} from "node-wot-protocol-coap"; 
 import {AssetResourceListener} from "node-wot-protocols";
 
 import * as http from "http";
