+++ packages/node-wot-protocol-coap/src/coap-server.ts
@@ -17,7 +17,10 @@
  * to copyright in this work will at all times remain with copyright holders.
  */
 
-/**
+
+
+
+ /**
  * CoAP Server based on coap by mcollina
  */
 
@@ -225,3 +228,4 @@ export default class CoapServer implements ProtocolServer {
     }
   }
 }
+
