+++ packages/node-wot-protocol-coap/tsconfig.json
