--- packages/node-wot-protocols-coap-client/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "node-wot-protocols-coap-client",
-  "version": "0.3.0-SNAPSHOT.2",
-  "files": [
-    "dist/"
-  ],
-  "description": "an coap client plugin for node-wot",
-  "main": "dist/coap-client-factory.js",
-  "typings": "dist/coap-client-factory.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "codestyle" : "standard --pretty"
-  },
-  "author": "the thingweb community",
-  "license": "W3C Software License",
-  "devDependencies": {
-    "@types/chai": "4.0.4",
-    "@types/node": "8.0.28",
-    "chai": "4.1.2",
-    "mocha": "3.5.3",
-    "mocha-typescript": "1.1.8",
-    "ts-node": "3.3.0",
-    "typescript": "2.5.2",
-    "typescript-standard": "0.3.30"
-  },
-  "dependencies": {
-    "coap": "0.20.0",
-    "node-wot-protocols": "0.3.0-SNAPSHOT.2"
-  }
-}
